[flutter_tools] make analyze once an integration test (#85336)
